Create a bean post processor to set the JAAS login module for Kafka
The JAAS login module to use is dependent on (but not the same as) the
value in KAFKA_SASL_MECHANISM.  The mechanism value something like
"PLAIN" or "SHA-SCRAM-512" while the login module value is a Kafka class
name.  We need a bean post processor to look at the value of
KAFKA_SASL_MECHANISM and then set the appropriate JAAS login module in
KafkaProperties.

This commit also updates how the truststore certificate is sent in to
KafkaProperties.  Clowder provides us with a PEM string so we have to
use (by way of KafkaProperties as a middle-man) the Kafka
ssl.truststore.certificates property (which just accepts a String
containing a certificate) rather than the prettier
ssl.truststore.location property which accepts a file path.